+++ tests/Mscc.GenerativeAI/GoogleAi_GeminiPro_Should.cs
@@ -924,6 +924,78 @@ namespace Test.Mscc.GenerativeAI
         }
         
         [Fact]
+        // Ref: https://ai.google.dev/gemini-api/docs/models/gemini-v2#search-tool
+        public async Task Generate_Content_with_Google_Search()
+        {
+            // Arrange
+            var prompt = "When is the next total solar eclipse in Mauritius?";
+            var genAi = new GoogleAI(fixture.ApiKey);
+            var model = genAi.GenerativeModel(Model.Gemini20FlashExperimental);
+            model.UseGoogleSearch = true;
+
+            // Act
+            var response = await model.GenerateContent(prompt);
+
+            // Assert
+            response.Should().NotBeNull();
+            response.Candidates.Should().NotBeNull().And.HaveCount(1);
+            response.Candidates![0].GroundingMetadata.Should().NotBeNull();
+            response.Candidates![0].GroundingMetadata!.SearchEntryPoint.Should().NotBeNull();
+            response.Candidates![0].GroundingMetadata!.WebSearchQueries.Should().NotBeNull();
+            output.WriteLine(string.Join(Environment.NewLine,
+                response.Candidates![0].Content!.Parts
+                    .Select(x => x.Text)
+//                    .Where(t => !string.IsNullOrEmpty(t))
+                    .ToArray()));
+            response.Candidates![0].GroundingMetadata!.GroundingChunks?
+                .ForEach(c => 
+                    output.WriteLine($"{c!.Web!.Title} - {c!.Web!.Uri}"));
+            output.WriteLine(string.Join(Environment.NewLine,
+                response.Candidates![0].GroundingMetadata!.WebSearchQueries!
+                    .Select(w => w)
+                    .ToArray()));
+            output.WriteLine(response.Candidates![0].GroundingMetadata!.SearchEntryPoint!.RenderedContent);
+        }
+        
+        [Fact]
+        // Ref: https://ai.google.dev/gemini-api/docs/models/gemini-v2#search-tool
+        public async Task Generate_Content_with_GoogleSearch_ResponseModalities()
+        {
+            // Arrange
+            var prompt = "When is the next total solar eclipse in Mauritius?";
+            var genAi = new GoogleAI(fixture.ApiKey);
+            var model = genAi.GenerativeModel(Model.Gemini20FlashExperimental);
+            model.UseGoogleSearch = true;
+
+            // Act
+            var response = await model.GenerateContent(prompt,
+                generationConfig: new GenerationConfig()
+                {
+                    ResponseModalities = [ ResponseModality.Text ]
+                });
+
+            // Assert
+            response.Should().NotBeNull();
+            response.Candidates.Should().NotBeNull().And.HaveCount(1);
+            response.Candidates![0].GroundingMetadata.Should().NotBeNull();
+            response.Candidates![0].GroundingMetadata!.SearchEntryPoint.Should().NotBeNull();
+            response.Candidates![0].GroundingMetadata!.WebSearchQueries.Should().NotBeNull();
+            output.WriteLine(string.Join(Environment.NewLine,
+                response.Candidates![0].Content!.Parts
+                    .Select(x => x.Text)
+//                    .Where(t => !string.IsNullOrEmpty(t))
+                    .ToArray()));
+            response.Candidates![0].GroundingMetadata!.GroundingChunks?
+                .ForEach(c => 
+                    output.WriteLine($"{c!.Web!.Title} - {c!.Web!.Uri}"));
+            output.WriteLine(string.Join(Environment.NewLine,
+                response.Candidates![0].GroundingMetadata!.WebSearchQueries!
+                    .Select(w => w)
+                    .ToArray()));
+            output.WriteLine(response.Candidates![0].GroundingMetadata!.SearchEntryPoint!.RenderedContent);
+        }
+        
+        [Fact]
         // Ref: https://ai.google.dev/docs/function_calling
         public async Task Function_Calling()
         {